requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
psycopg2-binary==2.9.10
PyJWT==1.7.1
sqlparse==0.5.1
